feat: Refactor project review page and implement judging interface with detailed evaluation criteria
